Creates new tasks
At this moment, the app get set of tasks and specific task, creates new
tasks once the validation checks whether the task exist or not.
The logic to update and delete is done. It is needed to check why it
does not interact with the browser for the time being.